fix: preserve active episode identity in page bridge
